fix add products

+++ controllers/DefaultController.php
@@ -9,6 +9,8 @@ error_reporting(1);
 class DefaultController extends CController
 {
     /** @var bool */
+    public $clearDir = false;
+    /** @var bool */
     public $zip = false;
     /** @var bool */
     public $adminNotify = false;
@@ -90,6 +92,9 @@ class DefaultController extends CController
     /** @var array */
     private $_status = array();
 
+    private $_im;
+
+
     public function init()
     {
         header('Content-type: text/plain; charset=windows-1251');
@@ -97,8 +102,8 @@ class DefaultController extends CController
 
     public function actionIndex($type = null, $mode = null, $filename = null, $dir = false)
     {
-        $IM = BaseImportModel::model($this->importClass);
-        $IM->module_name = $this->moduleUrl;
+        $this->_im = BaseImportModel::model($this->importClass);
+        $this->_im->module_name = $this->moduleUrl;
         if (empty($_GET) || !count($_GET)) {
             $url = explode('?', $_SERVER['REQUEST_URI']);
             $url = end($url);
@@ -124,7 +129,7 @@ class DefaultController extends CController
             case 'init':
                 if (Yii::app()->user->isGuest && !$this->register()) break;
 
-                CFileHelper::removeDirectory($this->baseDir);
+                if ($this->clearDir) CFileHelper::removeDirectory($this->baseDir);
 
                 $dir = $this->baseDir . date('Y-m-d_H-i_') . Yii::app()->user->id . DIRECTORY_SEPARATOR;
                 Yii::app()->user->setState('dir', $dir);
@@ -186,7 +191,7 @@ class DefaultController extends CController
             // Возвращаем такой же ответ
             case 'success':
                 if (Yii::app()->user->isGuest && !$this->register()) break;
-                $IM->clearBase(null, 'photo');
+                $this->_im->clearBase(null, 'photo');
                 $this->result('success');
 //                mail($this->email, '1C FINISHED ' . $type, "REQUEST: " . print_r($_REQUEST, 1) . "\r\nSERVER: " . print_r($_SERVER, 1));
                 break;
@@ -210,7 +215,6 @@ class DefaultController extends CController
 
     public function parseXml($path)
     {
-        $IM = BaseImportModel::model($this->importClass);
         $this->_status = Yii::app()->user->getState(Yii::app()->request->requestUri);
         $xml = simplexml_load_file($path);
 
@@ -227,16 +231,15 @@ class DefaultController extends CController
 
         if (isset($xml->Каталог))
             if ($xml->Каталог['СодержитТолькоИзменения'] == 'false')
-                $IM->clearBase($this->_timestamp, 'item');
+                $this->_im->clearBase($this->_timestamp, 'item');
 
         if (isset($xml->ПакетПредложений))
             if ($xml->ПакетПредложений['СодержитТолькоИзменения'] == 'false')
-                $IM->clearBase($this->_timestamp, 'offer');
+                $this->_im->clearBase($this->_timestamp, 'offer');
     }
 
     public function addData($data, $type)
     {
-        $IM = BaseImportModel::model($this->importClass);
         $i = 0;
         $count = $this->_status['count'][$type];
         if (!$count) $this->_status['count'][$type] = $count = count($data);
@@ -245,7 +248,7 @@ class DefaultController extends CController
         foreach ($data as $row) {
             if ($this->_status[$type] - 1 >= $i++) continue;
             $row = $this->trimAll($row);
-            $IM->createElement($row, $type, false, $this->_timestamp);
+            $this->_im->createElement($row, $type, false, $this->_timestamp);
 
             $this->_status[$type] = $i;
             Yii::app()->user->setState(Yii::app()->request->requestUri, $this->_status);
